requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
black==20.8b1
celery==5.0.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1
djangorestframework==3.12.1
flake8==3.8.4
kombu==5.0.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
redis==3.5.3
regex==2020.10.28
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
